requirements.txt
discord.py
yt-dlp
google-api-python-client
python-dotenv
flask
PyNaCl==1.5.0